basic sandbox for XSS Sim Project
This is a basic demo sandbox to simulate the project. Note that in real life, you might isolate the code using iframes or similar techniques. This is just a simple example of using an iframe to create a basic sandbox for simulating code execution. 

The sandbox attribute in the <iframe> tag is set to "allow-scripts," which allows the iframe to execute scripts